Reduced compiler warnings
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@111300 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/mime_magic/mime_magic.c
@@ -1425,7 +1425,7 @@ static int mget(union VALUETYPE *p, unsigned char *s,
 {
     long offset = m->offset;
 
-    if (offset + sizeof(union VALUETYPE) > nbytes)
+    if (offset + (long)sizeof(union VALUETYPE) > nbytes)
 		return 0;
 
     memcpy(p, s + offset, sizeof(union VALUETYPE));
@@ -1447,7 +1447,7 @@ static int mget(union VALUETYPE *p, unsigned char *s,
 			break;
 		}
 
-		if (offset + sizeof(union VALUETYPE) > nbytes)
+		if (offset + (long)sizeof(union VALUETYPE) > nbytes)
 			return 0;
 
 		memcpy(p, s + offset, sizeof(union VALUETYPE));
